requirements.txt
pytorch-lightning
git+https://github.com/openai/CLIP.git
omegaconf
transformers
einops
matplotlib
pillow